requirements.txt
gym==0.12.1
numpy==1.15.4
roboschool
torch==1.1.0
pyyaml==3.13
